pyproject.toml
# BEGIN DOTGIT-SYNC BLOCK MANAGED
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "dotgit-sync"
description = "Repository allowing to manage dotfile for git repos in an unified ways."
authors = [
  "Romain Deville <code@romaindeville.fr>",
]
maintainers = [
  "Romain Deville <code@romaindeville.fr>",
]
# BEGIN DOTGIT-SYNC BLOCK EXCLUDED CUSTOM_POETRY
version = "v0.3.1"
readme = "README.md"
keywords = []
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Topic :: Utilities",
    "Typing :: Typed",
]
packages = [
  { include = "dotgit_sync" },
]
include = [
  "dotgit_sync/schemas",
  "dotgit_sync/templates",
  "LICENSE*",
  "CHANGELOG.md",
  "AUTHORS"
]
homepage = "https://framagit.org/rdeville-public/programs/dotgit-sync/"
repository = "https://framagit.org/rdeville-public/programs/dotgit-sync/"
documentation = "https://framagit.org/rdeville-public/programs/dotgit-sync/"
# END DOTGIT-SYNC BLOCK EXCLUDED CUSTOM_POETRY

[tool.poetry.urls]
# BEGIN DOTGIT-SYNC BLOCK EXCLUDED CUSTOM_POETRY_URLS
issues = "https://framagit.org/rdeville-public/programs/dotgit-sync/-/issues"
# END DOTGIT-SYNC BLOCK EXCLUDED CUSTOM_POETRY_URLS

[tool.poetry.scripts]
# BEGIN DOTGIT-SYNC BLOCK EXCLUDED CUSTOM_POETRY_SCRIPTS
dotgit-sync = "dotgit_sync.main:main"
# END DOTGIT-SYNC BLOCK EXCLUDED CUSTOM_POETRY_SCRIPTS

[tool.poetry.group.dev.dependencies]
# BEGIN DOTGIT-SYNC BLOCK EXCLUDED CUSTOM_POETRY_DEV
yamllint = "^1.35.1"
ruff = "^0.4.8"
# END DOTGIT-SYNC BLOCK EXCLUDED CUSTOM_POETRY_DEV

[tool.poetry.group.test.dependencies]
# BEGIN DOTGIT-SYNC BLOCK EXCLUDED CUSTOM_POETRY_TEST
#
# END DOTGIT-SYNC BLOCK EXCLUDED CUSTOM_POETRY_TEST

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
# BEGIN DOTGIT-SYNC BLOCK EXCLUDED CUSTOM_POETRY_DOCS
#
# END DOTGIT-SYNC BLOCK EXCLUDED CUSTOM_POETRY_DOCS

[tool.poetry.dependencies]
python = "^3.12"
# BEGIN DOTGIT-SYNC BLOCK EXCLUDED CUSTOM_POETRY_DEP
gitpython = "^3.1.43"
jinja2 = "^3.1.4"
json5 = "^0.9.25"
pykwalify = "^1.8.0"
pylibmagic = "^0.5.0"
python-magic = "^0.4.27"
pyyaml = "^6.0.1"
requests = "^2.32.3"
# END DOTGIT-SYNC BLOCK EXCLUDED CUSTOM_POETRY_DEP

# END DOTGIT-SYNC BLOCK MANAGED
